requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
gunicorn>=22.0.0
reportlab>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.2
pytesseract>=0.3.10
Pillow>=10.0.0
psycopg2-binary>=2.9.0
